requirements.txt

---------------------------------------------

gunicorn
---------------------------------------------
mock
---------------------------------------------
Flask-Login
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask-WTF
beautifulsoup4
requests
psycopg2
flask-httpauth
passlib
PyJWT
flask-cors
